+++ java-api/src/main/java/com/db/grad/javaapi/service/UserService.java
@@ -1,0 +1,15 @@
+package com.db.grad.javaapi.service;
+
+import com.db.grad.javaapi.model.Dog;
+
+import com.db.grad.javaapi.model.User;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface UserService
+{
+    
+
+    public User getUserById(long uniqueId);
+}
